Organize routing functions (#1434)
